Reset source code to latest upstream state.

+++ llvm/include/llvm/ADT/STLExtras.h
@@ -1114,9 +1114,9 @@ public:
 
   iterator begin() const { return iterator(base, 0); }
   iterator end() const { return iterator(base, count); }
-  ReferenceT operator[](size_t index) const {
+  ReferenceT operator[](unsigned index) const {
     assert(index < size() && "invalid index for value range");
-    return DerivedT::dereference_iterator(base, static_cast<ptrdiff_t>(index));
+    return DerivedT::dereference_iterator(base, index);
   }
   ReferenceT front() const {
     assert(!empty() && "expected non-empty range");
